requirements.txt
# Package requirements
pathlib==3.9.13
PYPDF2==1.26.0
PyQt5.QtCore==5.15.7

# Development requirements
pytest==7.2.2
pytest-html==3.2.0
